PersonajePruebas

+++ AlgoBall/src/Modelo/Casillero.java
@@ -36,8 +36,8 @@ public class Casillero {
 		return pos_y;
 	}
 
-	public int distanciaA(Casillero objetivo) {
-		int resultado = (int) Math.sqrt(Math.pow(pos_x-objetivo.pos_x,2)+Math.pow(pos_y-objetivo.pos_y,2));
+	public double distanciaA(Casillero objetivo) {
+		double resultado = Math.sqrt(Math.pow(pos_x-objetivo.pos_x,2)+Math.pow(pos_y-objetivo.pos_y,2));
 		return resultado;
 	}
